requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
transformers==4.44.2
sentence-transformers==3.1.1
faiss-cpu==1.8.0
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
python-multipart==0.0.20